Downloads and reads ITR datasets

The ITR (informacoes trimestrais) is the quarterly reporting system of companies traded at B3. This function will access the CVM ftp and parse all available files according to user choices

get_itr_data(
  companies_cvm_codes = NULL,
  first_year = 2010,
  last_year = lubridate::year(Sys.Date()),
  type_docs = c("BPA", "BPP", "DRE"),
  type_format = c("con", "ind"),
  individual_dre_quarters = TRUE,
  clean_data = TRUE,
  use_memoise = FALSE,
  cache_folder = "gdfpd2_cache"
)

Arguments

companies_cvm_codes
Numeric CVM code of companies. IF set to NULL (default), will return data for all available companies.

first_year
First year of selected data

last_year
Last year of selected data

type_docs
Type of financial documents. E.g. c('DRE', 'BPA'). Definitions: '*' = fetch all docs, 'BPA' = Assets (ativos), 'BPP' = Liabilities (passivo), 'DRE' = income statement (demonstrativo de resultados), 'DFC_MD' = cash flow by direct method (fluxo de caixa pelo metodo direto), 'DFC_MI' = cash flow by indirect method (fluxo de caixa pelo metodo indireto), 'DMPL' = statement of changes in equity (mutacoes do patrimonio liquido), 'DVA' = value added report (desmonstrativo de valor agregado)

type_format
Type of format of document (con = consolidated, ind = individual). Default = c('con', 'ind')

individual_dre_quarters
Should the function calculate single quarters for companies: TRUE (default) or FALSE. This option only applies for the DRE. Be aware that the original data only includes accumulated quarters.

clean_data
Clean data or return raw data? See read_dfp|itr_csv() for details

use_memoise
Use memoise caching? If no (default), the function will read all .csv files. If yes, will use package memoise for caching results (execution speed increases significantly)

cache_folder
Path of cache folder to keep memoise and zip files

Value

A list of tibbles, separated by column GRUPO_DFP

Examples

if (FALSE)  # dontrun: keep cran check fast
df_itr <- get_itr_data() # fetches all available datasets